refactor(lib): move transifex
