+++ app/resources/v1/translate.js
@@ -1,6 +1,6 @@
 const fs = require('fs/promises')
 const logger = require('../../lib/logger.js')
-const getFromTransifex = require('../../../lib/transifex.js')
+const getFromTransifex = require('../../lib/transifex.js')
 
 async function getLocalTranslation (res, locale, resource) {
   const translationFile =
